+++ src/org/gearvrf/bullet/Geometry.java
@@ -19,7 +19,7 @@ import org.gearvrf.bullet.util.Vector3;
 
 public class Geometry {
 	
-	public int id = 0;
+	public long id = 0;
 	
 	
 	public Shape shape;
